+++ packages/contracts/hardhat.config.ts
@@ -6,9 +6,9 @@ dotenvConfig({ path: __dirname + "/.env" });
 const config: HardhatUserConfig = {
   solidity: "0.8.19",
   networks: {
-    goerli: {
-      url: `https://eth-goerli.g.alchemy.com/v2/${process.env.ALCHEMY_API_KEY}`,
-      accounts: [process.env.GOERLI_PRIVATE_KEY ?? ''],
+    sepolia: {
+      url: `https://eth-sepolia.g.alchemy.com/v2/${process.env.ALCHEMY_API_KEY}`,
+      accounts: [process.env.SEPOLIA_PRIVATE_KEY ?? ''],
     },
   },
   etherscan: {